server/requirements.txt
flask
flask_sqlalchemy
flask_cors
mysql-connector-python
python-decouple